Set version to 6.0.0
This is only to indicate that this line would only be compatible to
Angular 6. The 1.x line could be maintained for older versions if
necessary. Since all Angular packages have synchronized versions it
probably makes sense to also synchronize ngx-store as it is useful
in combination with Angular only (and v6 with this update).